refactor: Item 가격 자료형 변경

+++ wantedPayment/src/main/java/org/wantedpayment/trade/domain/dto/response/BuyHistoryResponse.java
@@ -4,11 +4,13 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import org.wantedpayment.trade.domain.entity.TradeStatus;
 
+import java.math.BigDecimal;
+
 @Getter
 @AllArgsConstructor
 public class BuyHistoryResponse {
     private Long tradeId;
-    private int tradePrice;
+    private BigDecimal tradePrice;
     private TradeStatus tradeStatus;
     private Long itemId;
     private String itemName;
